feat: add tweet functionality with CRUD operations and like support

+++ src/controllers/comment.controller.js
@@ -16,7 +16,7 @@ const getVideoComments = asyncHandler(async (req, res) => {
     }
 
     const comments = await Comment.find({ video: videoId })
-        .populate("commentedBy", "uesrname avatar")
+        .populate("commentedBy", "username avatar")
         .sort({ createdAt: -1 })
         .skip(skip)
         .limit(limit)
@@ -29,7 +29,7 @@ const getVideoComments = asyncHandler(async (req, res) => {
 
 })
 
-const addComment = asyncHandler(async (req, res) => {
+const addVideoComment = asyncHandler(async (req, res) => {
     const { videoId } = req.params;
     const userId = req.user._id;
     const { content } = req.body;
@@ -58,6 +58,7 @@ const addComment = asyncHandler(async (req, res) => {
 const updateComment = asyncHandler(async (req, res) => {
     const { commentId } = req.params
     const { content } = req.body
+    const userId = req.user._id
 
     const comment = await Comment.findByIdAndUpdate(
         commentId,
@@ -69,19 +70,28 @@ const updateComment = asyncHandler(async (req, res) => {
         throw new ApiError(404, "Comment not found")
     }
 
+    if (comment.owner.toString() !== userId.toString()) {
+        throw new ApiError(403, "You are not authorized to update/delete this comment")
+    }
+
     return res.status(200).json(
-        new APiResponse(200, null, "Comment updated successfully")
+        new APiResponse(200, comment, "Comment updated successfully")
     )
 
 })
 
 const deleteComment = asyncHandler(async (req, res) => {
     const { commentId } = req.params
+    const userId = req.user._id
 
     const comment = await Comment.findById(commentId)
 
     if (!comment) {
         throw new ApiError(404, "Comment not found")
+    }
+
+    if (!comment.owner.toString() !== userId.toString()) {
+        throw new ApiError(403, "your are not authorized to update/delete this comment")
     }
 
     await comment.deleteOne()
@@ -94,7 +104,7 @@ const deleteComment = asyncHandler(async (req, res) => {
 
 export {
     getVideoComments,
-    addComment,
+    addVideoComment,
     updateComment,
     deleteComment
 }
